Codeplex-777 Delta<T> can't be serialized with JSON.net
This is a partial fix for Codeplex-777, which asks that we make Delta<T>
work with the JsonSerializer.

This fix is only partial, because our implementation of GetMember and
SetMember don't perform conversions. So if you provide an int to a
property that wants a double, then it won't serialize that property. A
full fix for this requires some advanced use of JSON.Net's API and will be
published in a sample