Issue 1372 - ApiExplorer should add query parameters for complex FromUri types to the RelativePath
https://aspnetwebstack.codeplex.com/workitem/1372

The scope of this fix is for complex types that have FromUri attribute on
the parameter or on the class and for which all their properties are
convertible from a string. By design, it will skip complex parameters with
properties that are not convertible from string as it's not a common
scenario and complicates the fix.

For actions that have more than one complex type that matches the criteria
above, we will emit prefixes to avoid ambiguity, in any other case, we wont.