Fix Codeplex issue #1838
https://aspnetwebstack.codeplex.com/workitem/1838

Improve error message for the following case:
    @if (true)
    {
       <script> var a= '<script></script>' </script>
    }

The user code is wrong, and the script will be broken.
The browser will terminate the script at the first </script> inside
the string but will not report an error.

Razor catches it correctly, but the error message does not provide
enough clues to what went wrong.