Issue 1768 - OData V4 requires disabling ProxyCreation in EF https://aspnetwebstack.codeplex.com/workitem/1768
Reverted a change that made the internal cache of EdmStructuredObject
use a tuple of the property name and the IEdmStructruredType instead
of a tuple of the property name and the object type as the key for the
cache.

The issue was caused because the GetOrCreatePropertyGetter method
incorrectly cached a getter delegate using the EF dynamic proxy and
later on it tried to use that same getter with the actual type, which
resulted on a cast exception.

Reverting the cache to use the type instead of the IEdmStructuredType
will solve that problem as each type will have it's own getter. This is
also the behaviour the system exposed before model aliasing was
introduced.