Issue 1810 - V3 fix ODataProperties().Version should be calculated according to the spec. https://aspnetwebstack.codeplex.com/workitem/1810
The actual issue is that the version that was being calculated is the one for the
response and the codeplex issue describes the algorithm that calculates the request
version.

This change refactors the current v4 implementation for calculating the response by
moving the constants for the version headers to HttpRequestMessageProperties,
defining separate internal properties for the values of the OData-Version and
OData-MaxVersion and moving the logic for calculating the version to use in the
response to the ODataMediaTypeFormatter.