[Codeplex #1917]: User friendly message should be displayed if user requests service document with accept:application/xml [Codeplex #1939]: Should do some clean up in ODataMediaTypeFormatters.CreateApplicationXml
codeplex #1917 only mentions to clean up the service document in
application/xml. But, as a fact, ODL only supports Metadata document in
application/xml format. For other payload kinds  (for example, Property, EntityReferenceLink(s), Collection, Error)  are not supported any more as application/xml format in ODL.

codeplex #1939 is required to clean up all other not-supported payload kinds from application/xml format. The requirement is similar with the above and the fix is same.
So, this changeset merges them together. Thanks.