[Issue #2047] config.EnableQuerySupport() causes errors in Web API OData 5.2 (v3) https://aspnetwebstack.codeplex.com/workitem/2047
Given an app that has actions attributed with QueryableAttribute, if that
application also uses config.EnableQuerySupport() it will make the actions with
QueryableAttribute fail. config.EnableQuerySupport() adds an EnableQueryAttribute
to the collection of global filters, which causes the action to have two filters
applied: EnableQueryAttribute and QueryableAttribute.

The fix is to check for instances of QueryableAttribute applied to the action or
to the controller when the user calls config.EnableQuerySupport or
config.AddODataQueryFilter without providing an explicit IFilter instance.
In those cases we avoid applying the EnableQueryAttribute to the action globally.
(The filter on the action/controller will still be applied as expected).

If the user used a custom query provider to plug in the QueryableAttribute,
this issue won't apply when updating from 5.1 to a later version as the only
attribute that can be used in 5.1 is QueryableAttribute.

When a customer updates to 5.2-5.3 and provides a custom QueryableAttribute instance to
config.EnableQuerySupport, if he updates his custom QueryableAttribute instance to use
EnableQueryAttribute, he also needs to update all the actions that use
QueryableAttribute on his application to use EnableQueryAttribute to avoid a similar issue
with conflicting filters on such actions.